debug + js build (TODO: debug notifications)

+++ server/controllers/applications.coffee
@@ -1,6 +1,8 @@
-equest = require("request-json")
+request = require("request-json")
 fs = require('fs')
-{Application} = require '../models/application'
+slugify = require 'cozy-slug'
+
+Application = require '../models/application'
 {AppManager} = require '../lib/paas'
 {PermissionsManager} = require '../lib/permissions'
 {DescriptionManager} = require '../lib/description'
@@ -52,26 +54,31 @@ randomString = (length) ->
 
 # Save an app's icon in the DS
 saveIcon = (appli, callback = ->) ->
-    client = request.newClient "http://localhost:#{appli.port}/"
-    tmpName = "/tmp/icon_#{appli.slug}.png"
-    client.saveFile "icons/main_icon.png", tmpName, (err, res, body) ->
-        return callback err if err
-        appli.attachFile tmpName, name: 'icon.png', (err) ->
-            fs.unlink tmpName
+    if appli? and appli.port?
+        client = request.newClient "http://localhost:#{appli.port}/"
+        tmpName = "/tmp/icon_#{appli.slug}.png"
+        client.saveFile "icons/main_icon.png", tmpName, (err, res, body) ->
             return callback err if err
-            callback null
+            appli.attachFile tmpName, name: 'icon.png', (err) ->
+                fs.unlink tmpName
+                return callback err if err
+                callback null
+    else
+        callback new Error 'Appli cannot be reached'
+
 
 module.exports =
 
+
     # Load application corresponding to slug given in params
     loadApplication: (req, res, next, slug) ->
-        Application.all key: params.slug, (err, apps) ->
+        Application.all key: req.params.slug, (err, apps) ->
             if err
                 next err
             else if apps is null or apps.length is 0
                 res.send 404, error: 'Application not found'
             else
-                req.app
+                req.application = apps[0]
                 next()
 
 
@@ -100,11 +107,11 @@ module.exports =
     getMetaData: (req, res, next) ->
         metaDataManager = new DescriptionManager()
         metaDataManager.get req.body, (metaData) ->
-            res.send succes: true, app: metaData, 200
+            res.send success: true, app: metaData, 200
 
 
     read: (req, res, next) ->
-        Application.find params.id, (err, app) ->
+        Application.find req.params.id, (err, app) ->
             if err then send_error res, err
             else if app is null
                 send_error res, new Error('Application not found'), 404
@@ -113,19 +120,19 @@ module.exports =
 
 
     icon: (req, res, next) ->
-        if req.app._attachments?['icon.png']
-            return req.app.getFile('icon.png', (->)).pipe res
+        if req.application?._attachments?['icon.png']
+            return req.application.getFile('icon.png', (->)).pipe res
 
         # else, do the attaching (apps installed before)
         # FOR MIGRATION, REMOVE ME LATER
-        saveIcon req.app, (err) =>
+        saveIcon req.application, (err) =>
             if err
                 return fs.createReadStream('./client/app/assets/img/stopped.png').pipe res
-            req.app.getFile('icon.png', (->)).pipe res
+            req.application.getFile('icon.png', (->)).pipe res
 
 
     updatestoppable: (req, res, next) ->
-        Application.find params.id, (err, app) ->
+        Application.find req.params.id, (err, app) ->
             if err
                 send_error res, err
             else if app is null
@@ -164,7 +171,8 @@ module.exports =
 
                     res.send success: true, app: appli, 201
 
-                    console.info 'attempt to install app ' + JSON.stringify(appli)
+                    infos = JSON.stringify appli
+                    console.info "attempt to install app #{infos}"
                     manager = new AppManager()
                     manager.installApp appli, (err, result) ->
 
@@ -173,34 +181,40 @@ module.exports =
                             res.send_error_socket err
                             return
 
-                        appli.state = "installed"
-                        appli.port = result.drone.port
+                        if result.drone?
+                            appli.state = "installed"
+                            appli.port = result.drone.port
 
-                        console.info 'install succeeded on port ', appli.port
+                            msg = "install succeeded on port #{appli.port}"
+                            console.info msg
 
-                        saveIcon appli, (err) ->
-                            if err then console.log err.stack
-                            else console.info 'icon attached'
+                            saveIcon appli, (err) ->
+                                if err then console.log err.stack
+                                else console.info 'icon attached'
 
-                        appli.save (err) ->
-                            return res.send_error_socket err if err
-                            console.info 'saved port in db', appli.port
-                            manager.resetProxy (err) ->
-                                return send_error_socket err if err
-                                console.info 'proxy reset', appli.port
+                            appli.save (err) ->
+                                return res.send_error_socket err if err
+                                console.info 'saved port in db', appli.port
+                                manager.resetProxy (err) ->
+                                    return send_error_socket err if err
+                                    console.info 'proxy reset', appli.port
+                        else
+                            err = new Error "Controller has no " + \
+                                            "informations about #{appli.name}"
+                            return send_error_socket err if err
 
 
     # Remove app from 3 places :
-    # * haibu, application manager
+    # * haibu, application managerll
     # * proxy, cozy router
     # * database
     uninstall: (req, res, next) ->
-        req.body.slug = slugify req.body.name
-        manager = new AppManager
-        manager.uninstallApp req.app, (err, result) ->
-            return mark_broken res, req.app, err if err
+        req.body.slug = req.params.slug
+        manager = new AppManager()
+        manager.uninstallApp req.application, (err, result) ->
+            return mark_broken res, req.application, err if err
 
-            req.app.destroy (err) ->
+            req.application.destroy (err) ->
                 return send_error res, err if err
 
                 manager.resetProxy (err) ->
@@ -210,23 +224,24 @@ module.exports =
                         success: true
                         msg: 'Application succesfuly uninstalled'
 
+
     # Update an app :
     # * haibu, application manager
     # * proxy, cozy router
     # * database
     update: (req, res, next) ->
         manager = new AppManager()
-        if not req.app.password?
-            req.app.password = randomString 32
+        if not req.application.password?
+            req.application.password = randomString 32
 
-        manager.updateApp req.app, (err, result) ->
-            return mark_broken res, req.app, err if err
+        manager.updateApp req.application, (err, result) ->
+            return mark_broken res, req.application, err if err
 
-            req.app.state = "installed"
+            req.application.state = "installed"
             permissions = new PermissionsManager()
-            permissions.get req.app, (err, docTypes) ->
-                req.app.permissions = docTypes
-                req.app.save (err) ->
+            permissions.get req.application, (err, docTypes) ->
+                req.application.permissions = docTypes
+                req.application.save (err) ->
 
                     saveIcon appli, (err) ->
                         if err then console.log err.stack
@@ -235,45 +250,47 @@ module.exports =
                     return send_error res, err if err
 
                     manager.resetProxy (err) ->
-                        return mark_broken res, req.app, err if err
+                        return mark_broken res, req.application, err if err
 
                         res.send
                             success: true
                             msg: 'Application succesfuly updated'
 
+
     start: (req, res, next) ->
         manager = new AppManager
-        manager.start req.app, (err, result) ->
-            return mark_broken res, req.app, err if err
+        manager.start req.application, (err, result) ->
+            return mark_broken res, req.application, err if err
 
-            req.app.state = "installed"
-            req.app.port = result.drone.port
-            req.app.save (err) ->
+            req.application.state = "installed"
+            req.application.port = result.drone.port
+            req.application.save (err) ->
                 return send_error res, err if err
 
                 manager.resetProxy (err) ->
-                    return mark_broken res, req.app, err if err
+                    return mark_broken res, req.application, err if err
 
                     res.send
                         success: true
                         msg: 'Application running'
-                        app: req.app
+                        app: req.application
+
 
     stop: (req, res, next) ->
         manager = new AppManager
-        manager.stop req.app, (err, result) ->
-            return mark_broken res, req.app, err if err
+        manager.stop req.application, (err, result) ->
+            return mark_broken res, req.application, err if err
 
             data =
                 state: 'stopped'
                 port : 0
 
-            req.app.updateAttributes data, (err) ->
+            req.application.updateAttributes data, (err) ->
                 return send_error res, err if err
 
                 manager.resetProxy (err) ->
-                    return mark_broken res, req.app, err if err
+                    return mark_broken res, req.application, err if err
                     res.send
                         success: true
                         msg: 'Application stopped'
-                        app: req.app
+                        app: req.application
